requirements.txt
pillow==10.2.0
ultralytics==8.2.13
paho-mqtt==2.0.0
numpy==1.26.4
openvr==1.26.701
PyOpenGL==3.1.7
pygame==2.5.2